requirements.txt
amqp==2.3.2
celery==4.2.1
Django==1.8
django-extensions
django-registration==2.4
h5py==2.8.0
keras
matplotlib==2.2.3
numpy
pillow
pycairo
pymongo==3.7.2
redis
scipy
scikit-learn
tensorflow==1.4.1
theano
tqdm
uWSGI==2.0.17.1
